Add nondeterministic reduce that uses atomics (#4961)
This implementation of reduce uses atomics to accumulate the partial aggregates of each warp and the partial aggregates of each block. Compared to the default `run_to_run` reduce kernel, this launches only one kernel instead of two, although it does also require a call to `cudaMemsetAsync` to zero `d_out`.

On the H200 and the B200, the atomic reduce is ~20% faster for arrays of size 2^16. Larger array sizes are also measured to be faster, although the speedup is not as large.

Currently, this implementation only supports `cuda::std::plus<>` and will fall back to `run_to_run` for other operations. We plan on supporting more operators with time.